refactor(types): featureProps 타입 파일 분리
