+++ src/components/common/Modal.tsx
@@ -2,7 +2,7 @@ import { Fragment } from 'react';
 import { Dialog, Transition } from '@headlessui/react';
 import { X } from 'lucide-react';
 import { cn } from '@/lib/utils/cn';
-import { ModalProps } from "@/types/props";
+import { ModalProps } from "@/types/commonProps";
 
 export function Modal({
   isOpen,
