+++ src/components/common/Select.tsx
@@ -1,6 +1,6 @@
 import { forwardRef } from "react";
 import { cn } from "@/lib/utils/cn";
-import { SelectProps } from "@/types/props";
+import { SelectProps } from "@/types/commonProps";
 
 export const Select = forwardRef<HTMLSelectElement, SelectProps>(({
    className,
