+++ src/components/explore/SearchSection.tsx
@@ -1,4 +1,4 @@
-import {SearchSectionProps} from "@/types/props";
+import {SearchSectionProps} from "@/types/featureProps";
 import {SearchInput} from "@/components/common/SearchInput";
 
 export function SearchSection({ value, onChange }: SearchSectionProps) {
