+++ src/components/layout/Layout.tsx
@@ -1,7 +1,7 @@
 import { Header } from "./Header";
 import { Sidebar } from "./Sidebar";
 import { Footer } from "./Footer";
-import { LayoutProps } from "@/types/props";
+import { LayoutProps } from "@/types/commonProps";
 
 export const Layout = ({ children }: LayoutProps) => {
     return (
